--- src/conf/datamodule/EuroSAT_RGB_Samples.yaml
@@ -1,18 +1,0 @@
-_target_: datamodules.EuroSAT_RGB_Samples.EUsatrgbDM
-batch_size: 12
-dims:
-  height: 64
-  width: 64
-test_size: 0.2
-val_size: 0.2
-seed: 14
-normalize: True
-# num_workers: 1
-
-# dataset class
-Dataset:
-  _target_: datamodules.EuroSAT_RGB_Samples.EUsatrgbDS
-  dir: ${datamodule.dir}
-  im_height: 224
-  im_width: 224
-  im_channels: 3
