+++ src/conf/model/vitmae.yaml
@@ -3,10 +3,11 @@ wandb_plots: False
 # num_labels: 10
 
 net:
-  _target_: transformers.ViTForImageClassification.from_pretrained
-  pretrained_model_name_or_path: "google/vit-base-patch16-224"
-  num_labels: 10
-  ignore_mismatched_sizes: True
+#net:
+#  _target_: transformers.ViTForImageClassification.from_pretrained
+#  pretrained_model_name_or_path: "google/vit-base-patch16-224"
+#  num_labels: 10
+#  ignore_mismatched_sizes: True
 
 criterion:
   _target_: torch.nn.CrossEntropyLoss
