+++ src/models/base.py
@@ -2,7 +2,9 @@ from typing import Tuple
 
 import lightning as L
 import torch
+from torchmetrics.classification import MulticlassF1Score
 
+import wandb
 from datamodules.base import BaseDM
 
 """
@@ -117,3 +119,76 @@ class BaseModel(L.LightningModule):
 
     def logits_to_yhat(self, logits: torch.tensor) -> torch.tensor:
         return logits
+
+
+class MultiClass(BaseModel):
+    # base module for multi class problems
+    def __init__(
+        self, wandb_plots: bool = True, class_list: list[str] = None, **kwargs
+    ):
+        super().__init__(**kwargs)
+        self.f1fun = MulticlassF1Score(num_classes=10)
+        self.wandb_plots = wandb_plots
+        self.class_list = class_list
+
+    def logits_to_yhat(self, logits: torch.tensor) -> torch.tensor:
+        return torch.argmax(logits, dim=1)
+
+    def count_classes(self, y: torch.tensor, label: str) -> dict:
+        """Count num el in each class."""
+        counter = {}
+        for idx, classn in enumerate(self.class_list):
+            counter[f"{classn}"] = torch.sum(y == idx)
+        return counter
+
+    def end_of_epoch_metrics(
+        self,
+        logits: list[torch.tensor],
+        y: list[torch.tensor],
+        yhat: list[torch.tensor],
+        label: str,
+    ) -> dict:
+        yhat = torch.cat(self.yhat_all)
+        y = torch.cat(self.y_all)
+        f1 = self.f1fun(yhat, y)
+        counter = self.count_classes(yhat, label)
+        metrics = {}
+        metrics[f"{label}_fl"] = f1
+        if label == "test":
+            acc = self.accuracy(yhat, y)
+            logits = torch.cat(self.logits_all)
+            loss = self.criterion(logits, y)
+            metrics[f"{label}_acc"] = acc
+            metrics[f"{label}_loss"] = loss
+        self.logger.log_metrics(metrics)
+        if self.wandb_plots:
+            confmat = wandb.plot.confusion_matrix(
+                y_true=y,
+                preds=yhat,
+                # y_true=y.numpy(),
+                # preds=yhat.numpy(),
+                class_names=self.class_list,
+                title=f"{label}_confmat",
+            )
+            wandb.log({f"{label}_conf_mat": confmat})
+            data = [[key, int(counter[key].numpy())] for key in counter]
+            table = wandb.Table(data=data, columns=["class", "count"])
+            wandb.log(
+                {
+                    f"{label}_class_counter": wandb.plot.bar(
+                        table, "class", "count", title=f"{label}_class_counter"
+                    )
+                }
+            )
+
+    def on_validation_epoch_end(self) -> None:
+        self.end_of_epoch_metrics(self.logits_all, self.y_all, self.yhat_all, "val")
+
+    def on_test_epoch_end(self) -> None:
+        self.end_of_epoch_metrics(self.logits_all, self.y_all, self.yhat_all, "test")
+
+    def on_train_epoch_end(self) -> None:
+        if self.epoch_counter % 10 == 0:
+            self.end_of_epoch_metrics(
+                self.logits_all, self.y_all, self.yhat_all, "train"
+            )
